Endret litt struktur og lagt til simpel validering pa tlfnr
