+++ package.json
@@ -36,26 +36,26 @@
     "moment": "^2.24.0",
     "moment-timezone": "^0.5.26",
     "mustache-express": "^1.3.0",
-    "nav-frontend-alertstriper": "^3.0.6",
-    "nav-frontend-alertstriper-style": "^2.0.5",
+    "nav-frontend-alertstriper": "^3.0.7",
+    "nav-frontend-alertstriper-style": "^2.0.6",
     "nav-frontend-chevron": "^1.0.10",
     "nav-frontend-chevron-style": "^0.3.5",
-    "nav-frontend-core": "^4.0.10",
-    "nav-frontend-ikoner-assets": "^1.0.2",
+    "nav-frontend-core": "^4.0.11",
+    "nav-frontend-ikoner-assets": "^1.0.3",
     "nav-frontend-js-utils": "^1.0.7",
     "nav-frontend-knapper": "^1.0.36",
     "nav-frontend-knapper-style": "^0.3.31",
     "nav-frontend-lenker": "^1.0.31",
     "nav-frontend-lenker-style": "^0.2.23",
     "nav-frontend-paneler": "^1.0.20",
-    "nav-frontend-paneler-style": "^0.3.16",
+    "nav-frontend-paneler-style": "^0.3.17",
     "nav-frontend-skjema": "^1.0.75",
     "nav-frontend-skjema-style": "^1.0.47",
     "nav-frontend-spinner": "^1.0.19",
     "nav-frontend-spinner-style": "^0.2.5",
     "nav-frontend-tekstomrade": "^2.0.35",
-    "nav-frontend-typografi": "^2.0.15",
-    "nav-frontend-typografi-style": "^1.0.16",
+    "nav-frontend-typografi": "^2.0.16",
+    "nav-frontend-typografi-style": "^1.0.17",
     "nav-frontend-veileder": "^2.0.12",
     "nav-frontend-veileder-style": "^2.0.9",
     "nav-frontend-veilederpanel": "^0.1.15",
