+++ src/App.less
@@ -28,13 +28,6 @@ body {
   }
 }
 
-.spørsmålsgruppe {
-  padding-bottom: 2rem;
-  &:last-child {
-    padding-bottom: 0;
-  }
-}
-
 .panel {
   width: 800px;
 
@@ -44,20 +37,5 @@ body {
 
   @media @mobile {
     width: 100%;
-  }
-}
-
-.gridSvar {
-  .skjema__fieldset {
-    .inputPanelGruppe__inner {
-      display: grid;
-      grid-template-columns: repeat(2, 1fr);
-      grid-template-rows: min-content;
-      grid-gap: 1rem;
-
-      @media @mobile {
-        grid-template-columns: 1fr;
-      }
-    }
   }
 }
