+++ src/index.less
@@ -11,7 +11,6 @@
 @import "~nav-frontend-typografi-style";
 
 @import "./App.less";
-@import "./components/søknad/Søknad.less";
+@import "søknad/Søknad.less";
 @import "./components/feilside/Feilside.less";
-@import "components/språkvelger/Språkvelger.less";
-@import "./inngangsvilkår/personopplysninger/Personopplysninger.less";
+@import "./components/språkvelger/Språkvelger.less";
