+++ src/søknad/Søknad.less
@@ -1,10 +1,13 @@
+.søknadsdialog {
+}
+
 .personopplysninger {
   & > section {
     margin-bottom: 2rem;
   }
 
-  .typo-element {
-    padding-top: 1rem;
+  & > .typo-normal {
+    padding-bottom: 1.5rem;
   }
 }
